--- app/src/main/java/com/example/spotifycloneapp/SpotifyApiService.kt
@@ -1,13 +1,0 @@
-package com.example.spotifycloneapp
-
-import retrofit2.Call
-import retrofit2.http.GET
-
-interface SpotifyApiService {
-    @GET("popularRadios.json")
-    fun getPopularRadios(): Call<List<Radio>>
-
-    @GET("locationRadios.json")
-    fun getLocationRadios(): Call<List<Radio>>
-
-}
