+++ app/src/main/java/com/example/spotifycloneapp/data/remote/RadioServiceProvider.kt
@@ -1,12 +1,15 @@
-package com.example.spotifycloneapp
+package com.example.spotifycloneapp.data.remote
 
+import com.example.spotifycloneapp.BuildConfig
 import retrofit2.Retrofit
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
 
 class RadioServiceProvider {
 
     private val retrofit = Retrofit.Builder()
         .addConverterFactory(GsonConverterFactory.create())
+        .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
         .baseUrl(BuildConfig.SERVER_URL)
         .build()
