+++ app/src/main/java/com/example/spotifycloneapp/data/remote/SpotifyApiService.kt
@@ -1,0 +1,13 @@
+package com.example.spotifycloneapp.data.remote
+
+import io.reactivex.Single
+import retrofit2.http.GET
+
+interface SpotifyApiService {
+    @GET("popularRadios.json")
+    fun getPopularRadios(): Single<List<Radio>>
+
+    @GET("locationRadios.json")
+    fun getLocationRadios(): Single<List<Radio>>
+
+}
